1) Moved some floats in Particle structure that don’t have a small dynamic range of values from FLOAT to float to save memory.  2) Removed sinkid variable from Particle structure and replaced with flag to indicate if inside sink or not.

+++ src/Common/Dust.cpp
@@ -1071,7 +1071,7 @@ void DustSemiImplictForces<ndim, ParticleType, StoppingTime, Kernel>::ComputeDra
       gsound = neiblist[j].sound ;
       grho = neiblist[j].rho ;
       drho = parti.rho ;
-      parti.sound = max(parti.sound, gsound) ;
+      parti.sound = max((FLOAT) parti.sound, gsound) ;
       parti.div_v = max(parti.div_v, sqrt(DotProduct(dv,dv,ndim))/parti.h);
     }
